refactor: move openai client to internal package
Moves the OpenAI client implementation to the internal/openai package to improve code organization.
