+++ cmd/root.go
@@ -1,95 +1,16 @@
 package cmd
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"os/exec"
 	"strings"
 
+	"github.com/sestrella/autocommitmsg/internal/openai"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
-
-type OpenAIClient struct {
-	httpClient http.Client
-	baseUrl    string
-}
-
-type OpenAITransport struct {
-	base   http.RoundTripper
-	apiKey string
-}
-
-type CreateResponse struct {
-	Model    string    `json:"model"`
-	Messages []Message `json:"messages"`
-}
-
-type Message struct {
-	Role    string `json:"role"`
-	Content string `json:"content"`
-}
-
-type CreatedResponse struct {
-	Choices []Choice `json:"choices"`
-}
-
-type Choice struct {
-	Message ChoiceMessage `json:"message"`
-}
-
-type ChoiceMessage struct {
-	Content string `json:"content"`
-}
-
-func (transport OpenAITransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", transport.apiKey))
-	req.Header.Set("Content-Type", "application/json")
-	return transport.base.RoundTrip(req)
-}
-
-func newOpenAIClient(baseUrl string, apiKey string) OpenAIClient {
-	transport := OpenAITransport{
-		base:   http.DefaultTransport,
-		apiKey: apiKey,
-	}
-	httpClient := http.Client{Transport: transport}
-	return OpenAIClient{httpClient, baseUrl}
-}
-
-func (client OpenAIClient) createChatCompletion(model string, input []Message) (*CreatedResponse, error) {
-	body, err := json.Marshal(CreateResponse{model, input})
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequest("POST", fmt.Sprintf("%s/chat/completions", client.baseUrl), bytes.NewBuffer(body))
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := client.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	if res.StatusCode != 200 {
-		return nil, fmt.Errorf("response status code is not 200: %+v", res)
-	}
-
-	var data CreatedResponse
-	err = json.NewDecoder(res.Body).Decode(&data)
-	if err != nil {
-		return nil, err
-	}
-
-	return &data, nil
-}
 
 var cfgFile string
 
@@ -155,8 +76,8 @@ var rootCmd = &cobra.Command{
 			cobra.CheckErr("base-url cannot be empty")
 		}
 
-		client := newOpenAIClient(baseUrl, apiKey)
-		res, err := client.createChatCompletion(model, []Message{
+		client := openai.NewClient(baseUrl, apiKey)
+		res, err := client.CreateChatCompletion(model, []openai.Message{
 			{
 				Role:    "developer",
 				Content: "You are an assistant that writes concise, conventional commit messages based on the provided git diff. Return the commit message without any quotes.",
